Model/requirements.txt
numpy
pandas
scikit-learn
xgboost
pyyaml
pydantic
shap
pillow
tensorflow
open-clip-torch
ultralytics
faiss-cpu
